added catalog, mibo products and services apis & UI.
added forget password UI & APIs
added notifications module
added weight and calories compare

+++ app/src/main/java/life/mibo/android/models/weight/CompareWeightResponse.kt
@@ -20,7 +20,7 @@ class CompareWeightResponse(data: List<Data?>?) :
         @SerializedName("MemberID")
         var memberID: Int?,
         @SerializedName("Weight")
-        var weight: Int?,
+        var weight: Double?,
         @SerializedName("Weight Unit")
         var weightUnit: String?
     )
